+++ pkgs/watcher/lib/src/directory_watcher.dart
@@ -7,7 +7,7 @@ import 'dart:io';
 import '../watcher.dart';
 import 'custom_watcher_factory.dart';
 import 'directory_watcher/linux.dart';
-import 'directory_watcher/mac_os.dart';
+import 'directory_watcher/macos.dart';
 import 'directory_watcher/windows_resubscribable_watcher.dart';
 
 /// Watches the contents of a directory and emits [WatchEvent]s when something
@@ -54,7 +54,7 @@ abstract class DirectoryWatcher implements Watcher {
       );
       if (customWatcher != null) return customWatcher;
       if (Platform.isLinux) return LinuxDirectoryWatcher(directory);
-      if (Platform.isMacOS) return MacOSDirectoryWatcher(directory);
+      if (Platform.isMacOS) return MacosDirectoryWatcher(directory);
       if (Platform.isWindows) {
         return WindowsDirectoryWatcher(directory,
             runInIsolate: runInIsolateOnWindows);