+++ pkgs/watcher/CHANGELOG.md
@@ -37,8 +37,9 @@
   issues: tracking failure following subdirectory move, incorrect events when
   there are changes in a recently-moved subdirectory, incorrect events due to
   various situations involving subdirectory moves.
-- Bug fix: with `DirectoryWatcher` on MacOS, fix events for changes in new
-  directories: don't emit duplicate ADD, don't emit MODIFY without ADD.
+- Bug fix: new `DirectoryWatcher` implementation on MacOS that fixes various
+  issues including duplicate events for changes in new directories, incorrect
+  events when close together directory renames have overlapping names.
 - Bug fix: with `FileWatcher` on MacOS, a modify event was sometimes reported if
   the file was created immediately before the watcher was created. Now, if the
   file exists when the watcher is created then this modify event is not sent.